+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Add support for connecting to HDFS ([#263]).
 - Add support for Hive 3.1.3 ([#243]).
 - PVCs for data storage, cpu and memory limits are now configurable ([#270]).
+- Add temporary attribute to support using ClusterIP instead of NodePort service type ([#272]).
 
 ### Changed
 
@@ -31,6 +32,7 @@ All notable changes to this project will be documented in this file.
 [#244]: https://github.com/stackabletech/trino-operator/pull/244
 [#263]: https://github.com/stackabletech/trino-operator/pull/263
 [#270]: https://github.com/stackabletech/trino-operator/pull/270
+[#272]: https://github.com/stackabletech/trino-operator/pull/272
 
 ## [0.4.0] - 2022-06-30